InfoTable modification + removed most of the HQL queries made in the application side

+++ eclipse_ws/biobank2/src/edu/ualberta/med/biobank/model/ClinicStudyInfo.java
@@ -9,9 +9,9 @@ public class ClinicStudyInfo {
 
     public String studyShortName;
 
-    public int patients;
+    public long patients;
 
-    public Long patientVisits;
+    public long patientVisits;
 
     public void performDoubleClick() {
         SessionManager.getInstance().openViewForm(Study.class,
